+++ src/assets/ElementMapBundle.php
@@ -4,7 +4,7 @@
  * @copyright Copyright Charlie Development
  */
 
-namespace charliedev\elementmap\resources;
+namespace charliedev\elementmap\assets;
 
 use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
@@ -13,7 +13,7 @@ class ElementMapBundle extends AssetBundle
 {
 	public function init()
 	{
-		$this->sourcePath = '@charliedev/elementmap/resources/dist';
+		$this->sourcePath = '@charliedev/elementmap/assets/dist';
 
 		$this->depends = [
 			CpAsset::class,